catalytic-computing-production/requirements.txt
# Core dependencies
numpy==1.24.3
scipy==1.11.4
networkx==3.1
pandas==2.0.3

# GPU acceleration
cupy-cuda12x==12.2.0
numba==0.58.1

# Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# API and Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-benchmark==4.0.0
hypothesis==6.92.1

# Development tools
black==23.11.0
mypy==1.7.1
pylint==3.0.3
pre-commit==3.5.0

# Deployment
gunicorn==21.2.0
docker==6.1.3
kubernetes==28.1.0

# Caching and Performance
redis==5.0.1
cachetools==5.3.2
joblib==1.3.2

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4